+++ NEWS.md
@@ -10,3 +10,10 @@
 
 * README.md now contains instructions for installing from CRAN as well as from GitHub.
 
+# 2019-02-06 - bnpsd 1.0.2.9000
+
+* Added rbnpsd(noFixed) option to redraw loci that were drawn fixed for a single allele.
+These loci are not polymorphic so they would normally not be considered in analyses.
+
+* Added function fixed_loci to test for fixed loci within rbnpsd.
+